refactor: extract player zone controllers
