+++ lib/widgets/player_zone_widget.dart
@@ -1,4 +1,3 @@
 export 'player_zone/player_zone_core.dart';
-export 'player_zone/player_zone_animator.dart';
 export 'player_zone/player_zone_overlay.dart';
 export 'player_zone/player_zone_action_panel.dart';
